v1.0.0 Azure & Advice launch
Major update, 1.0.0 because older updates no longer are functional and a major update has been added:

The api now runs on Azure app, officially making this a microservice!

Due to ALOT of changes to make the transition towards Azure, the API feature has been removed. this will be resolved as further minor updates on the new version will drop.

the container and image has been fully rewritten, as the api doesn't use a database, docker-compose has been removed.

A new application should be build for the model training part, however as for DDD, the training part is NOT something that is used by personal in the end, i will ask David for feedback on how to approach this

!REMEMBER TO TURN OFF THE AZURE WEB APP UNLESS YOU WANT YOUR CREDITS GONE BECAUSE OF BOTS! THIS WILL BE RESOLVED WHEN A API KEY IS APPLIED!